refactor: switch currency representation to long cents

+++ src/main/java/io/github/shadrahman/fingraph/model/Account.java
@@ -6,7 +6,7 @@ public record Account(
         String id,
         String userId,
         String name,
-        Double balance,
+        Long balance,
         List<Transaction> history
 ) {
 }
