+++ src/main/java/io/github/shadrahman/fingraph/model/Transaction.java
@@ -2,7 +2,7 @@ package io.github.shadrahman.fingraph.model;
 
 public record Transaction(
         String id,
-        Double amount,
+        Long amount,
         String description,
         Category category,
         boolean isSubscription
